UPSTREAM:CP ath12k:wifi: Fix bridge vdev crash and peer association regression
Recent changes to punc_bitmap introduced a kernel crash during bridge
vdev start on branch 4.0. Additionally, a firmware assert occurs
during bridge peer association due to incorrect num_partner_links being
sent.

This is caused by the change of ahsta->links_map from u32 to u16, which
is insufficient for bridge link IDs 23 and 24.
This patch addresses both issues.

Fixes: Ica174584267 : (UPSTREAM:CP: wifi: ath12k: Cleanup patch for
       links_map access)
       I53ba36be4e9 : (PKG upgrade rebase Progressive DFS/AGDFS)

+++ backports-6.1-9a0dddfb3/drivers/net/wireless/ath/ath12k/core.h
@@ -831,7 +831,7 @@ struct ath12k_sta {
 	struct ath12k_link_sta deflink;
 	struct ath12k_link_sta __rcu *link[ATH12K_NUM_MAX_LINKS];
 	/* indicates bitmap of link sta created in FW */
-	u16 links_map;
+	u32 links_map;
 	u8 assoc_link_id;
 	u16 ml_peer_id;
 	u8 num_peer;
